pyproject.toml
[project]
name = "studybuddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.9.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
